+++ src/test/suite/testFiles/propValue.tsx
@@ -1,0 +1,8 @@
+import { useTranslations } from "@vocab/react";
+import translations from "./.vocab";
+import React from "react";
+
+const MyComponent = () => {
+  const { t } = useTranslations(translations);
+  return <div foo="bar">{t("Already extracted")}</div>;
+};